Revert "feat: Add support for Action Items (#1596)"
This reverts commit a71db3bdc0ef16ed035d742dc3c7bd336e5ec016.
