+++ public/locales/en.json
@@ -286,40 +286,6 @@
     "eventCreated": "Congratulations! The Event is created.",
     "talawaApiUnavailable": "Talawa-API service is unavailable. Is it running? Check your network connectivity too."
   },
-  "organizationActionItems": {
-    "actionItemCategory": "Action Item Category",
-    "actionItemDetails": "Action Item Details",
-    "assignee": "Assignee",
-    "assigner": "Assigner",
-    "assignmentDate": "Assignment Date",
-    "active": "Active",
-    "clearFilters": "Clear Filters",
-    "close": "Close",
-    "completed": "Completed",
-    "completionDate": "Completion Date",
-    "createActionItem": "Create",
-    "deleteActionItem": "Delete Action Item",
-    "deleteActionItemMsg": "Do you want to remove this action item?",
-    "details": "Details",
-    "dueDate": "Due Date",
-    "earliest": "Earliest",
-    "editActionItem": "Edit Action Item",
-    "isCompleted": "Completed",
-    "latest": "Latest",
-    "no": "No",
-    "noActionItems": "No Action Items",
-    "options": "Options",
-    "preCompletionNotes": "Pre Completion Notes",
-    "postCompletionNotes": "Post Completion Notes",
-    "selectActionItemCategory": "Select an action item category",
-    "selectAssignee": "Select an assignee",
-    "status": "Status",
-    "successfulCreation": "Action Item created successfully",
-    "successfulUpdation": "Action Item updated successfully",
-    "successfulDeletion": "Action Item deleted successfully",
-    "title": "Action Items",
-    "yes": "Yes"
-  },
   "eventListCard": {
     "location": "Location",
     "on": "On",
