+++ public/locales/fr.json
@@ -281,40 +281,6 @@
     "eventCreated": "Toutes nos félicitations! L'événement est créé.",
     "talawaApiUnavailable": "Le service Talawa-API n'est pas disponible. Est-il en cours d'exécution ? Vérifiez également votre connectivité réseau."
   },
-  "organizationActionItems": {
-    "actionItemCategory": "Catégorie de l'élément d'action",
-    "actionItemDetails": "Détails de l'élément d'action",
-    "assignee": "Attributaire",
-    "assigner": "Assignateur",
-    "assignmentDate": "Date d'attribution",
-    "active": "Actif",
-    "clearFilters": "Effacer les filtres",
-    "close": "Fermer",
-    "completed": "Terminé",
-    "completionDate": "Date d'achèvement",
-    "createActionItem": "Créer",
-    "deleteActionItem": "Supprimer l'élément d'action",
-    "deleteActionItemMsg": "Voulez-vous supprimer cet élément d'action?",
-    "details": "Détails",
-    "dueDate": "Date d'échéance",
-    "earliest": "Le plus tôt",
-    "editActionItem": "Modifier l'élément d'action",
-    "isCompleted": "Terminé",
-    "latest": "Le plus récent",
-    "no": "Non",
-    "noActionItems": "Pas d'éléments d'action",
-    "options": "Options",
-    "preCompletionNotes": "Notes avant achèvement",
-    "postCompletionNotes": "Notes après achèvement",
-    "selectActionItemCategory": "Sélectionnez une catégorie d'élément d'action",
-    "selectAssignee": "Sélectionnez un attributaire",
-    "status": "Statut",
-    "successfulCreation": "Élément d'action créé avec succès",
-    "successfulUpdation": "Élément d'action mis à jour avec succès",
-    "successfulDeletion": "Élément d'action supprimé avec succès",
-    "title": "Éléments d'action",
-    "yes": "Oui"
-  },
   "eventListCard": {
     "location": "Emplacement",
     "on": "Sur",
