+++ public/locales/hi.json
@@ -281,41 +281,6 @@
     "eventCreated": "बधाई हो! इवेंट बनाया गया है।",
     "talawaApiUnavailable": "तलवा-एपीआई सेवा उपलब्ध नहीं है। क्या यह चल रहा है? अपनी नेटवर्क कनेक्टिविटी भी जांचें।"
   },
-  "organizationActionItems": {
-    "actionItemCategory": "कार्य आइटम श्रेणी",
-    "actionItemDetails": "कार्य आइटम विवरण",
-    "assignee": "सौंपने वाला",
-    "assigner": "निर्धारक",
-    "assignmentDate": "सौंपने की तारीख",
-    "active": "सक्रिय",
-    "clearFilters": "फ़िल्टर हटाएं",
-    "close": "बंद करें",
-    "completed": "पूरा हुआ",
-    "completionDate": "समापन तिथि",
-    "createActionItem": "बनाएं",
-    "deleteActionItem": "कार्य आइटम हटाएं",
-    "deleteActionItemMsg": "क्या आप इस कार्य आइटम को हटाना चाहते हैं?",
-    "details": "विवरण",
-    "dueDate": "नियत तारीख",
-    "earliest": "सबसे पहले",
-    "editActionItem": "कार्य आइटम संपादित करें",
-    "isCompleted": "पूरा हुआ",
-    "latest": "नवीनतम",
-    "no": "नहीं",
-    "noActionItems": "कोई कार्य आइटम नहीं",
-    "options": "विकल्प",
-    "preCompletionNotes": "समापन से पहले के नोट्स",
-    "postCompletionNotes": "समापन के बाद के नोट्स",
-    "searchByName": "नाम से खोजें",
-    "selectActionItemCategory": "कार्य आइटम श्रेणी का चयन करें",
-    "selectAssignee": "सौंपने वाला चुनें",
-    "status": "स्थिति",
-    "successfulCreation": "कार्य आइटम सफलतापूर्वक बनाया गया",
-    "successfulUpdation": "कार्य आइटम सफलतापूर्वक अपडेट किया गया",
-    "successfulDeletion": "कार्य आइटम सफलतापूर्वक हटाया गया",
-    "title": "कार्य आइटम",
-    "yes": "हाँ"
-  },
   "eventListCard": {
     "location": "स्थान",
     "on": "पर",
