+++ public/locales/sp.json
@@ -281,40 +281,6 @@
     "eventCreated": "¡Felicidades! Se crea el Evento.",
     "talawaApiUnavailable": "El servicio Talawa-API no está disponible. ¿Está funcionando? Compruebe también la conectividad de su red."
   },
-  "organizationActionItems": {
-    "actionItemCategory": "Categoría del ítem de acción",
-    "actionItemDetails": "Detalles del ítem de acción",
-    "assignee": "Asignado",
-    "assigner": "Asignador",
-    "assignmentDate": "Fecha de asignación",
-    "active": "Activo",
-    "clearFilters": "Borrar filtros",
-    "close": "Cerrar",
-    "completed": "Completado",
-    "completionDate": "Fecha de finalización",
-    "createActionItem": "Crear",
-    "deleteActionItem": "Eliminar ítem de acción",
-    "deleteActionItemMsg": "¿Desea eliminar este ítem de acción?",
-    "details": "Detalles",
-    "dueDate": "Fecha de vencimiento",
-    "earliest": "Lo más temprano",
-    "editActionItem": "Editar ítem de acción",
-    "isCompleted": "Completado",
-    "latest": "Lo más reciente",
-    "no": "No",
-    "noActionItems": "No hay ítems de acción",
-    "options": "Opciones",
-    "preCompletionNotes": "Notas previas a la finalización",
-    "postCompletionNotes": "Notas posteriores a la finalización",
-    "selectActionItemCategory": "Seleccione una categoría de ítem de acción",
-    "selectAssignee": "Seleccione un asignado",
-    "status": "Estado",
-    "successfulCreation": "Ítem de acción creado con éxito",
-    "successfulUpdation": "Ítem de acción actualizado con éxito",
-    "successfulDeletion": "Ítem de acción eliminado con éxito",
-    "title": "Ítems de acción",
-    "yes": "Sí"
-  },
   "eventListCard": {
     "location": "Ubicación",
     "on": "En",
