+++ public/locales/zh.json
@@ -281,41 +281,6 @@
     "eventCreated": "恭喜！事件已創建。",
     "talawaApiUnavailable": "服務不可用。它在運行嗎？還要檢查您的網絡連接。"
   },
-  "organizationActionItems": {
-    "actionItemCategory": "行动项目类别",
-    "actionItemDetails": "行动项目详情",
-    "assignee": "受让人",
-    "assigner": "分配者",
-    "assignmentDate": "分配日期",
-    "active": "活跃",
-    "clearFilters": "清除筛选器",
-    "close": "关闭",
-    "completed": "已完成",
-    "completionDate": "完成日期",
-    "createActionItem": "创建",
-    "deleteActionItem": "删除行动项目",
-    "deleteActionItemMsg": "您是否要删除此行动项目？",
-    "details": "详情",
-    "dueDate": "到期日",
-    "earliest": "最早",
-    "editActionItem": "编辑行动项目",
-    "isCompleted": "已完成",
-    "latest": "最新",
-    "no": "否",
-    "noActionItems": "无行动项目",
-    "options": "选项",
-    "preCompletionNotes": "完成前的笔记",
-    "postCompletionNotes": "完成后的笔记",
-    "searchByName": "按名称搜索",
-    "selectActionItemCategory": "选择行动项目类别",
-    "selectAssignee": "选择受让人",
-    "status": "状态",
-    "successfulCreation": "行动项目创建成功",
-    "successfulUpdation": "行动项目更新成功",
-    "successfulDeletion": "行动项目删除成功",
-    "title": "行动项目",
-    "yes": "是"
-  },
   "eventListCard": {
     "location": "地點",
     "on": "上",
