+++ src/GraphQl/Mutations/mutations.ts
@@ -231,6 +231,33 @@ export const DELETE_ORGANIZATION_MUTATION = gql`
   }
 `;
 
+// to create an action item category
+
+export const CREATE_ACTION_ITEM_CATEGORY_MUTATION = gql`
+  mutation CreateActionItemCategory($name: String!, $organizationId: ID!) {
+    createActionItemCategory(name: $name, organizationId: $organizationId) {
+      _id
+    }
+  }
+`;
+
+// to update an action item category
+
+export const UPDATE_ACTION_ITEM_CATEGORY_MUTATION = gql`
+  mutation UpdateActionItemCategory(
+    $actionItemCategoryId: ID!
+    $name: String
+    $isDisabled: Boolean
+  ) {
+    updateActionItemCategory(
+      id: $actionItemCategoryId
+      data: { name: $name, isDisabled: $isDisabled }
+    ) {
+      _id
+    }
+  }
+`;
+
 // to create the event by any organization
 
 export const CREATE_EVENT_MUTATION = gql`
@@ -280,6 +307,7 @@ export const DELETE_EVENT_MUTATION = gql`
 `;
 
 // to remove an admin from an organization
+
 export const REMOVE_ADMIN_MUTATION = gql`
   mutation RemoveAdmin($orgid: ID!, $userid: ID!) {
     removeAdmin(data: { organizationId: $orgid, userId: $userid }) {
@@ -289,6 +317,7 @@ export const REMOVE_ADMIN_MUTATION = gql`
 `;
 
 // to Remove member from an organization
+
 export const REMOVE_MEMBER_MUTATION = gql`
   mutation RemoveMember($orgid: ID!, $userid: ID!) {
     removeMember(data: { organizationId: $orgid, userId: $userid }) {
@@ -298,6 +327,7 @@ export const REMOVE_MEMBER_MUTATION = gql`
 `;
 
 // to add the admin
+
 export const ADD_ADMIN_MUTATION = gql`
   mutation CreateAdmin($orgid: ID!, $userid: ID!) {
     createAdmin(data: { organizationId: $orgid, userId: $userid }) {
@@ -575,15 +605,6 @@ export const REGISTER_EVENT = gql`
     }
   }
 `;
-
-// Create and Update Action Item Categories
-export { CREATE_ACTION_ITEM_CATEGORY_MUTATION } from './ActionItemCategoryMutations';
-export { UPDATE_ACTION_ITEM_CATEGORY_MUTATION } from './ActionItemCategoryMutations';
-
-// Create, Update and Delete Action Items
-export { CREATE_ACTION_ITEM_MUTATION } from './ActionItemMutations';
-export { UPDATE_ACTION_ITEM_MUTATION } from './ActionItemMutations';
-export { DELETE_ACTION_ITEM_MUTATION } from './ActionItemMutations';
 
 // Changes the role of a event in an organization and add and remove the event from the organization
 export { ADD_EVENT_ATTENDEE } from './EventAttendeeMutations';
