+++ src/GraphQl/Queries/Queries.ts
@@ -2,6 +2,7 @@ import gql from 'graphql-tag';
 
 //Query List
 // Check Auth
+
 export const CHECK_AUTH = gql`
   query {
     checkAuth {
@@ -85,6 +86,18 @@ export const ORGANIZATION_CONNECTION_LIST = gql`
         sortingCode
         state
       }
+    }
+  }
+`;
+
+// Query to get the action item category list
+
+export const ACTION_ITEM_CATEGORY_LIST = gql`
+  query ActionItemCategoriesByOrganization($organizationId: ID!) {
+    actionItemCategoriesByOrganization(organizationId: $organizationId) {
+      _id
+      name
+      isDisabled
     }
   }
 `;
@@ -330,6 +343,7 @@ export const ORGANIZATIONS_LIST = gql`
 `;
 
 // Query to take the Members of a particular organization
+
 export const MEMBERS_LIST = gql`
   query Organizations($id: ID!) {
     organizations(id: $id) {
@@ -543,6 +557,7 @@ export const ORGANIZATION_DONATION_CONNECTION_LIST = gql`
 `;
 
 // to take the list of the admins of a particular
+
 export const ADMIN_LIST = gql`
   query Organizations($id: ID!) {
     organizations(id: $id) {
@@ -560,6 +575,7 @@ export const ADMIN_LIST = gql`
 `;
 
 // to take the membership request
+
 export const MEMBERSHIP_REQUEST = gql`
   query Organizations($id: ID!) {
     organizations(id: $id) {
@@ -576,12 +592,6 @@ export const MEMBERSHIP_REQUEST = gql`
     }
   }
 `;
-
-// get the list of Action Item Categories
-export { ACTION_ITEM_CATEGORY_LIST } from './ActionItemCategoryQueries';
-
-// get the list of Action Items
-export { ACTION_ITEM_LIST } from './ActionItemQueries';
 
 // to take the list of the blocked users
 export { PLUGIN_GET } from './PlugInQueries';
