--- src/components/ActionItemsContainer/ActionItemsContainer.module.css
@@ -1,4 +1,0 @@
-.actionItemStatusBadge {
-  width: 5.5rem;
-  margin-left: 1.1rem;
-}
