--- src/screens/OrganizationActionItems/OrganizationActionItems.module.css
@@ -1,174 +1,0 @@
-.actionItemsContainer {
-  height: 90vh;
-}
-
-.actionItemModal {
-  max-width: 80vw;
-  margin-top: 2vh;
-  margin-left: 13vw;
-}
-
-.btnsContainer {
-  display: flex;
-  gap: 10px;
-}
-
-.btnsContainer .btnsBlock {
-  display: flex;
-  gap: 10px;
-}
-
-.btnsContainer button {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.container {
-  min-height: 100vh;
-}
-
-.createActionItemButton {
-  position: absolute;
-  top: 1.3rem;
-  right: 2rem;
-}
-
-.datediv {
-  display: flex;
-  flex-direction: row;
-}
-
-.datebox {
-  width: 90%;
-  border-radius: 7px;
-  outline: none;
-  box-shadow: none;
-  padding-top: 2px;
-  padding-bottom: 2px;
-  padding-right: 5px;
-  padding-left: 5px;
-  margin-right: 5px;
-  margin-left: 5px;
-}
-
-.dropdown {
-  display: block;
-}
-
-.dropdownToggle {
-  margin-bottom: 0;
-  display: flex;
-}
-
-.dropdownModalToggle {
-  width: 50%;
-}
-
-.errorIcon {
-  transform: scale(1.5);
-  color: var(--bs-danger);
-  margin-bottom: 1rem;
-}
-
-.greenregbtn {
-  margin: 1rem 0 0;
-  margin-top: 15px;
-  border: 1px solid var(--bs-gray-300);
-  box-shadow: 0 2px 2px var(--bs-gray-300);
-  padding: 10px 10px;
-  border-radius: 5px;
-  background-color: var(--bs-primary);
-  width: 100%;
-  font-size: 16px;
-  color: var(--bs-white);
-  outline: none;
-  font-weight: 600;
-  cursor: pointer;
-  transition:
-    transform 0.2s,
-    box-shadow 0.2s;
-  width: 100%;
-}
-
-hr {
-  border: none;
-  height: 1px;
-  background-color: var(--bs-gray-500);
-  margin: 1rem;
-}
-
-.iconContainer {
-  display: flex;
-  justify-content: flex-end;
-}
-.icon {
-  margin: 1px;
-}
-
-.message {
-  margin-top: 25%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-
-.organizationActionItemsContainer h2 {
-  margin: 0.6rem 0;
-}
-
-.preview {
-  display: flex;
-  flex-direction: row;
-  font-weight: 900;
-  font-size: 16px;
-  color: rgb(80, 80, 80);
-}
-
-.removeFilterIcon {
-  cursor: pointer;
-}
-
-.searchForm {
-  display: inline;
-}
-
-.titlemodal {
-  color: var(--bs-gray-600);
-  font-weight: 600;
-  font-size: 20px;
-  margin-bottom: 20px;
-  padding-bottom: 5px;
-  border-bottom: 3px solid var(--bs-primary);
-  width: 65%;
-}
-
-.view {
-  margin-left: 2%;
-  font-weight: 600;
-  font-size: 16px;
-  color: var(--bs-gray-600);
-}
-
-@media (max-width: 767px) {
-  .btnsContainer {
-    margin-bottom: 0;
-    display: flex;
-    flex-direction: column;
-  }
-
-  .btnsContainer .btnsBlock .dropdownToggle {
-    flex-grow: 1;
-  }
-
-  .btnsContainer button {
-    width: 100%;
-  }
-
-  .createActionItemButton {
-    position: absolute;
-    top: 1rem;
-    right: 2rem;
-  }
-}
