+++ src/state/reducers/routesReducer.test.ts
@@ -14,7 +14,6 @@ describe('Testing Routes reducer', () => {
         { name: 'Dashboard', url: '/orgdash/id=undefined' },
         { name: 'People', url: '/orgpeople/id=undefined' },
         { name: 'Events', url: '/orgevents/id=undefined' },
-        { name: 'Action Items', url: '/orgactionitems/id=undefined' },
         { name: 'Posts', url: '/orgpost/id=undefined' },
         {
           name: 'Block/Unblock',
@@ -50,11 +49,6 @@ describe('Testing Routes reducer', () => {
           name: 'Events',
           comp_id: 'orgevents',
           component: 'OrganizationEvents',
-        },
-        {
-          name: 'Action Items',
-          comp_id: 'orgactionitems',
-          component: 'OrganizationActionItems',
         },
         { name: 'Posts', comp_id: 'orgpost', component: 'OrgPost' },
         { name: 'Block/Unblock', comp_id: 'blockuser', component: 'BlockUser' },
@@ -94,7 +88,6 @@ describe('Testing Routes reducer', () => {
         { name: 'Dashboard', url: '/orgdash/id=undefined' },
         { name: 'People', url: '/orgpeople/id=undefined' },
         { name: 'Events', url: '/orgevents/id=undefined' },
-        { name: 'Action Items', url: '/orgactionitems/id=undefined' },
         { name: 'Posts', url: '/orgpost/id=undefined' },
         { name: 'Block/Unblock', url: '/blockuser/id=undefined' },
         { name: 'Advertisement', url: '/orgads/id=undefined' },
@@ -128,11 +121,6 @@ describe('Testing Routes reducer', () => {
           name: 'Events',
           comp_id: 'orgevents',
           component: 'OrganizationEvents',
-        },
-        {
-          name: 'Action Items',
-          comp_id: 'orgactionitems',
-          component: 'OrganizationActionItems',
         },
         { name: 'Posts', comp_id: 'orgpost', component: 'OrgPost' },
         { name: 'Block/Unblock', comp_id: 'blockuser', component: 'BlockUser' },
@@ -172,7 +160,6 @@ describe('Testing Routes reducer', () => {
         { name: 'Dashboard', url: '/orgdash/id=undefined' },
         { name: 'People', url: '/orgpeople/id=undefined' },
         { name: 'Events', url: '/orgevents/id=undefined' },
-        { name: 'Action Items', url: '/orgactionitems/id=undefined' },
         { name: 'Posts', url: '/orgpost/id=undefined' },
         {
           name: 'Block/Unblock',
@@ -212,11 +199,7 @@ describe('Testing Routes reducer', () => {
           comp_id: 'orgevents',
           component: 'OrganizationEvents',
         },
-        {
-          name: 'Action Items',
-          comp_id: 'orgactionitems',
-          component: 'OrganizationActionItems',
-        },
+
         { name: 'Posts', comp_id: 'orgpost', component: 'OrgPost' },
         { name: 'Block/Unblock', comp_id: 'blockuser', component: 'BlockUser' },
         {
