+++ src/state/reducers/routesReducer.ts
@@ -67,11 +67,6 @@ const components: ComponentType[] = [
   { name: 'Dashboard', comp_id: 'orgdash', component: 'OrganizationDashboard' },
   { name: 'People', comp_id: 'orgpeople', component: 'OrganizationPeople' },
   { name: 'Events', comp_id: 'orgevents', component: 'OrganizationEvents' },
-  {
-    name: 'Action Items',
-    comp_id: 'orgactionitems',
-    component: 'OrganizationActionItems',
-  },
   { name: 'Posts', comp_id: 'orgpost', component: 'OrgPost' },
   { name: 'Block/Unblock', comp_id: 'blockuser', component: 'BlockUser' },
   { name: 'Advertisement', comp_id: 'orgads', component: 'Advertisements' },
@@ -88,6 +83,7 @@ const components: ComponentType[] = [
       },
     ],
   },
+
   { name: 'Settings', comp_id: 'orgsetting', component: 'OrgSettings' },
   { name: '', comp_id: 'member', component: 'MemberDetail' },
 ];
